+++ packages/sdk/src/models/table.ts
@@ -840,6 +840,7 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
             type: MutationTypes.SET_MULTIPLE_RECORDS_CELL_VALUES,
             tableId: this.id,
             records: recordsWithCellValuesByFieldId,
+            opts: {parseDateCellValueInColumnTimeZone: true},
         });
     }
     /**
@@ -1430,6 +1431,7 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
             type: MutationTypes.CREATE_MULTIPLE_RECORDS,
             tableId: this.id,
             records: recordsToCreate,
+            opts: {parseDateCellValueInColumnTimeZone: true},
         });
 
         return recordsToCreate.map(record => record.id);