+++ packages/sdk/test/models/table.test.ts
@@ -1143,6 +1143,9 @@ describe('Table', () => {
                             cellValuesByFieldId: {},
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1162,6 +1165,9 @@ describe('Table', () => {
                             cellValuesByFieldId: {},
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1187,6 +1193,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1212,6 +1221,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1227,6 +1239,9 @@ describe('Table', () => {
                     type: MutationTypes.CREATE_MULTIPLE_RECORDS,
                     tableId: 'tblDesignProjects',
                     records: [],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1246,6 +1261,9 @@ describe('Table', () => {
                             cellValuesByFieldId: {},
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1275,6 +1293,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1304,6 +1325,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -1972,6 +1996,9 @@ describe('Table', () => {
                     type: MutationTypes.SET_MULTIPLE_RECORDS_CELL_VALUES,
                     tableId: 'tblDesignProjects',
                     records: [{id: 'recA', cellValuesByFieldId: {}}],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2017,6 +2044,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2041,6 +2071,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2067,6 +2100,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2093,6 +2129,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2120,6 +2159,9 @@ describe('Table', () => {
                     type: MutationTypes.SET_MULTIPLE_RECORDS_CELL_VALUES,
                     tableId: 'tblDesignProjects',
                     records: [],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2133,6 +2175,9 @@ describe('Table', () => {
                     type: MutationTypes.SET_MULTIPLE_RECORDS_CELL_VALUES,
                     tableId: 'tblDesignProjects',
                     records: [{id: 'recA', cellValuesByFieldId: {}}],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2198,6 +2243,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );
@@ -2241,6 +2289,9 @@ describe('Table', () => {
                             },
                         },
                     ],
+                    opts: {
+                        parseDateCellValueInColumnTimeZone: true,
+                    },
                 },
                 {holdForMs: 100},
             );